+++ core/src/ten_manager/designer_frontend/public/locales/zh-CN/common.json
@@ -8,6 +8,7 @@
       "autoLayout": "自动布局",
       "openAppFolder": "打开应用文件夹",
       "file": "文件",
+      "app": "应用",
       "help": "帮助",
       "about": "关于",
       "language": "语言",
@@ -18,7 +19,8 @@
       "theme.dark": "暗黑",
       "theme.system": "系统",
       "start": "启动",
-      "preferences": "偏好设置"
+      "preferences": "偏好设置",
+      "installAll": "安装所有"
     },
     "officialSite": "官方网站",
     "poweredBy": "技术支持",
