+++ core/src/ten_manager/designer_frontend/src/components/AppBar/Menu/EditMenu.tsx
